Revert frontend changes
