+++ client/web/src/enterprise/campaigns/apply/ChangesetSpecNode.story.tsx
@@ -1,10 +1,10 @@
 import { storiesOf } from '@storybook/react'
 import React from 'react'
-import { ChangesetSpecNode } from './ChangesetApplyPreviewNode'
+import { ChangesetSpecNode } from './ChangesetSpecNode'
 import { visibleChangesetSpecStories } from './VisibleChangesetSpecNode.story'
 import { hiddenChangesetSpecStories } from './HiddenChangesetSpecNode.story'
 import { of } from 'rxjs'
-import { EnterpriseWebStory } from '../../../components/EnterpriseWebStory'
+import { EnterpriseWebStory } from '../../components/EnterpriseWebStory'
 
 const { add } = storiesOf('web/campaigns/apply/ChangesetSpecNode', module).addDecorator(story => (
     <div className="p-3 container web-content changeset-spec-list__grid">{story()}</div>
