+++ client/web/src/enterprise/campaigns/apply/HiddenChangesetSpecNode.tsx
@@ -1,18 +1,21 @@
 import React from 'react'
-import { ChangesetApplyPreviewFields } from '../../../../graphql-operations'
+import { HiddenChangesetSpecFields, ChangesetSpecType } from '../../../graphql-operations'
 import InfoCircleOutlineIcon from 'mdi-react/InfoCircleOutlineIcon'
 import { ChangesetSpecAction } from './ChangesetSpecAction'
 
 export interface HiddenChangesetSpecNodeProps {
-    node: ChangesetApplyPreviewFields
+    node: HiddenChangesetSpecFields
 }
 
 export const HiddenChangesetSpecNode: React.FunctionComponent<HiddenChangesetSpecNodeProps> = ({ node }) => (
     <>
         <span className="d-none d-sm-block" />
-        <ChangesetSpecAction node={node} className="hidden-changeset-spec-node__action" />
+        <ChangesetSpecAction spec={node} className="hidden-changeset-spec-node__action" />
         <div className="d-flex flex-column hidden-changeset-spec-node__information">
-            <h3 className="text-muted">Changeset in a private repository</h3>
+            <h3 className="text-muted">
+                {node.type === ChangesetSpecType.EXISTING && <>Import changeset from a private repository</>}
+                {node.type === ChangesetSpecType.BRANCH && <>Create changeset in a private repository</>}
+            </h3>
             <span className="text-danger">
                 No action will be taken on apply.{' '}
                 <InfoCircleOutlineIcon
