+++ client/web/src/enterprise/campaigns/apply/VisibleChangesetSpecNode.story.tsx
@@ -2,10 +2,10 @@ import { storiesOf } from '@storybook/react'
 import React from 'react'
 import { VisibleChangesetSpecNode } from './VisibleChangesetSpecNode'
 import { addDays } from 'date-fns'
-import { VisibleChangesetSpecFields, ChangesetSpecType, Scalars } from '../../../../graphql-operations'
+import { VisibleChangesetSpecFields, ChangesetSpecType, Scalars } from '../../../graphql-operations'
 import { of } from 'rxjs'
-import { EnterpriseWebStory } from '../../../components/EnterpriseWebStory'
-import { ChangesetSpecOperation } from '../../../../../../shared/src/graphql-operations'
+import { EnterpriseWebStory } from '../../components/EnterpriseWebStory'
+import { ChangesetSpecOperation } from '../../../../../shared/src/graphql-operations'
 
 const { add } = storiesOf('web/campaigns/apply/VisibleChangesetSpecNode', module).addDecorator(story => (
     <div className="p-3 container web-content changeset-spec-list__grid">{story()}</div>
