+++ client/web/src/enterprise/campaigns/apply/backend.ts
@@ -1,10 +1,15 @@
-import { diffStatFields } from '../../../backend/diff'
+import { diffStatFields, fileDiffFields } from '../../../backend/diff'
 import { gql, dataOrThrowErrors } from '../../../../../shared/src/graphql/graphql'
 import {
     Scalars,
     CampaignSpecFields,
     CampaignSpecByIDVariables,
     CampaignSpecByIDResult,
+    CampaignSpecChangesetSpecsResult,
+    CampaignSpecChangesetSpecsVariables,
+    ChangesetSpecFileDiffsVariables,
+    ChangesetSpecFileDiffsResult,
+    ChangesetSpecFileDiffsFields,
     CreateCampaignVariables,
     CreateCampaignResult,
     ApplyCampaignResult,
@@ -14,7 +19,7 @@ import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators'
 import { requestGraphQL } from '../../../backend/graphql'
 
-const viewerCampaignsCodeHostsFragment = gql`
+export const viewerCampaignsCodeHostsFragment = gql`
     fragment ViewerCampaignsCodeHostsFields on CampaignsCodeHostConnection {
         totalCount
         nodes {
@@ -31,7 +36,7 @@ const supersedingCampaignSpecFragment = gql`
     }
 `
 
-const campaignSpecFragment = gql`
+export const campaignSpecFragment = gql`
     fragment CampaignSpecFields on CampaignSpec {
         id
         description {
@@ -96,6 +101,193 @@ export const fetchCampaignSpecById = (campaignSpec: Scalars['ID']): Observable<C
                 throw new Error(`The given ID is a ${node.__typename}, not a CampaignSpec`)
             }
             return node
+        })
+    )
+
+export const changesetSpecFieldsFragment = gql`
+    fragment ChangesetSpecFields on ChangesetSpec {
+        __typename
+        ... on HiddenChangesetSpec {
+            ...HiddenChangesetSpecFields
+        }
+        ... on VisibleChangesetSpec {
+            ...VisibleChangesetSpecFields
+        }
+    }
+
+    fragment CommonChangesetSpecFields on ChangesetSpec {
+        expiresAt
+        type
+        operations
+        delta {
+            titleChanged
+        }
+        changeset {
+            __typename
+            id
+            ... on ExternalChangeset {
+                title
+            }
+        }
+    }
+
+    fragment HiddenChangesetSpecFields on HiddenChangesetSpec {
+        __typename
+        id
+        ...CommonChangesetSpecFields
+    }
+
+    fragment VisibleChangesetSpecFields on VisibleChangesetSpec {
+        __typename
+        id
+        ...CommonChangesetSpecFields
+        description {
+            __typename
+            ...ExistingChangesetReferenceFields
+            ...GitBranchChangesetDescriptionFields
+        }
+    }
+
+    fragment ExistingChangesetReferenceFields on ExistingChangesetReference {
+        baseRepository {
+            name
+            url
+        }
+        externalID
+    }
+
+    fragment GitBranchChangesetDescriptionFields on GitBranchChangesetDescription {
+        baseRepository {
+            name
+            url
+        }
+        title
+        published
+        body
+        commits {
+            author {
+                avatarURL
+                email
+                displayName
+                user {
+                    username
+                    displayName
+                    url
+                }
+            }
+            subject
+            body
+        }
+        baseRef
+        headRef
+        diffStat {
+            ...DiffStatFields
+        }
+    }
+
+    ${diffStatFields}
+`
+
+export const queryChangesetSpecs = ({
+    campaignSpec,
+    first,
+    after,
+}: CampaignSpecChangesetSpecsVariables): Observable<
+    (CampaignSpecChangesetSpecsResult['node'] & { __typename: 'CampaignSpec' })['changesetSpecs']
+> =>
+    requestGraphQL<CampaignSpecChangesetSpecsResult, CampaignSpecChangesetSpecsVariables>(
+        gql`
+            query CampaignSpecChangesetSpecs($campaignSpec: ID!, $first: Int, $after: String) {
+                node(id: $campaignSpec) {
+                    __typename
+                    ... on CampaignSpec {
+                        changesetSpecs(first: $first, after: $after) {
+                            totalCount
+                            pageInfo {
+                                endCursor
+                                hasNextPage
+                            }
+                            nodes {
+                                ...ChangesetSpecFields
+                            }
+                        }
+                    }
+                }
+            }
+
+            ${changesetSpecFieldsFragment}
+        `,
+        { campaignSpec, first, after }
+    ).pipe(
+        map(dataOrThrowErrors),
+        map(({ node }) => {
+            if (!node) {
+                throw new Error(`CampaignSpec with ID ${campaignSpec} does not exist`)
+            }
+            if (node.__typename !== 'CampaignSpec') {
+                throw new Error(`The given ID is a ${node.__typename}, not a CampaignSpec`)
+            }
+            return node.changesetSpecs
+        })
+    )
+
+export const changesetSpecFileDiffsFields = gql`
+    fragment ChangesetSpecFileDiffsFields on VisibleChangesetSpec {
+        description {
+            __typename
+            ... on GitBranchChangesetDescription {
+                diff {
+                    fileDiffs(first: $first, after: $after) {
+                        nodes {
+                            ...FileDiffFields
+                        }
+                        totalCount
+                        pageInfo {
+                            hasNextPage
+                            endCursor
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    ${fileDiffFields}
+`
+
+export const queryChangesetSpecFileDiffs = ({
+    changesetSpec,
+    first,
+    after,
+    isLightTheme,
+}: ChangesetSpecFileDiffsVariables): Observable<
+    (ChangesetSpecFileDiffsFields['description'] & { __typename: 'GitBranchChangesetDescription' })['diff']
+> =>
+    requestGraphQL<ChangesetSpecFileDiffsResult, ChangesetSpecFileDiffsVariables>(
+        gql`
+            query ChangesetSpecFileDiffs($changesetSpec: ID!, $first: Int, $after: String, $isLightTheme: Boolean!) {
+                node(id: $changesetSpec) {
+                    __typename
+                    ...ChangesetSpecFileDiffsFields
+                }
+            }
+
+            ${changesetSpecFileDiffsFields}
+        `,
+        { changesetSpec, first, after, isLightTheme }
+    ).pipe(
+        map(dataOrThrowErrors),
+        map(({ node }) => {
+            if (!node) {
+                throw new Error(`ChangesetSpec with ID ${changesetSpec} does not exist`)
+            }
+            if (node.__typename !== 'VisibleChangesetSpec') {
+                throw new Error(`The given ID is a ${node.__typename}, not a VisibleChangesetSpec`)
+            }
+            if (node.description.__typename !== 'GitBranchChangesetDescription') {
+                throw new Error('The given ChangesetSpec is no GitBranchChangesetDescription')
+            }
+            return node.description.diff
         })
     )
